simpletrafficracer/try_new_flow/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.7.1            # via tb-nightly, tensorflow-gpu, tf-agents-nightly
astor==0.8.0              # via tensorflow-gpu
astroid==2.2.5            # via pylint
autopep8==1.4.4
backcall==0.1.0           # via ipython
cloudpickle==1.1.1        # via tfp-nightly
cycler==0.10.0            # via matplotlib
decorator==4.4.0          # via ipython, tfp-nightly, traitlets
gast==0.2.2               # via tensorflow-gpu
gin-config==0.1.3         # via tf-agents-nightly
google-pasta==0.1.7       # via tensorflow-gpu
grpcio==1.22.0            # via tb-nightly, tensorflow-gpu
h5py==2.9.0
ipython-genutils==0.2.0   # via traitlets
ipython==7.7.0
isort==4.3.21             # via pylint
jedi==0.14.1              # via ipython
keras-applications==1.0.8  # via tensorflow-gpu
keras-preprocessing==1.1.0  # via tensorflow-gpu
kiwisolver==1.1.0         # via matplotlib
lazy-object-proxy==1.4.1  # via astroid
markdown==3.1.1           # via tb-nightly
matplotlib==3.1.1
mccabe==0.6.1             # via pylint
mock==3.0.5               # via tf-agents-nightly
numpy==1.17.0
opencv-python==4.1.0.25
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==6.1.0
prompt-toolkit==2.0.9     # via ipython
protobuf==3.9.0           # via tb-nightly, tensorflow-gpu
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.5.0        # via autopep8
pydot==1.4.1
pygame==1.9.6
pygments==2.4.2           # via ipython
pylint==2.3.1
pyparsing==2.4.2          # via matplotlib, pydot
python-dateutil==2.8.0    # via matplotlib
pyyaml==5.1.1
six==1.12.0               # via absl-py, astroid, cycler, gin-config, grpcio, h5py, keras-preprocessing, mock, prompt-toolkit, protobuf, python-dateutil, tb-nightly, tensorflow-gpu, tf-agents-nightly, tfp-nightly, traitlets
tb-nightly==1.14.0a20190603  # via tensorflow-gpu
tensorflow-gpu==2.0.0b1
termcolor==1.1.0          # via tensorflow-gpu
tf-agents-nightly==0.2.0.dev20190716
tf-estimator-nightly==1.14.0.dev2019060501  # via tensorflow-gpu
tfp-nightly==0.8.0.dev20190730  # via tf-agents-nightly
tqdm==4.32.2
traitlets==4.3.2          # via ipython
typed-ast==1.4.0          # via astroid
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.15.5          # via tb-nightly
wheel==0.33.4             # via tb-nightly, tensorflow-gpu
wrapt==1.11.2             # via astroid, tensorflow-gpu

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via ipython, kiwisolver, markdown, protobuf, tb-nightly
